requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.32
botocore==1.27.32
cloudpickle==2.1.0
cycler==0.11.0
dask==2022.2.0
Django==3.2.13
django-dotenv==1.4.2
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dlib==19.24.0
easydict==1.9
fsspec==2022.5.0
gunicorn==20.1.0
imageio==2.3.0
jmespath==1.0.1
kiwisolver==1.4.3
locket==1.0.0
matplotlib==3.5.1
networkx==2.6.3
pandas==0.23.4
partd==1.2.0
Pillow==6.2.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
scikit-image==0.14.2
scikit-learn==0.19.2
scipy==1.1.0
sqlparse==0.4.2
toolz==0.11.2
tqdm==4.24.0
wincertstore==0.2
tzdata==2022.1